+++ src/daemon/mod.rs
@@ -2,6 +2,7 @@ pub mod api;
 pub mod bootstrap;
 pub mod cli;
 pub mod core;
+pub mod env_grabber;
 pub mod error;
 pub mod executor;
 pub mod logs;